Add files via upload
Assignment1: Use a checkbox to control the readonly property of a textbox. If checkbox is check make the textbox readonly and If the check is unchecked make the textbox editable
Assignment2: Create 2 groups of checkboxes and display all the values of the checkboxes checked by the user. Same scenario as we did with radio input types in the class.